[skip ci] update ConfigTable
