+++ app/src/main/java/me/singleneuron/qn_kernel/tlb/ConfigTable.kt
@@ -1,13 +1,47 @@
 package me.singleneuron.qn_kernel.tlb
 
-object ConfigTable {
-    var isTim: Boolean = false
+import nil.nadph.qnotified.util.Utils
 
-    fun <T> getConfig(className: String?) : T {
-        if(isTim) {
-            return TIMConfigTable.getConfig(className) as T
+object ConfigTable {
+
+    private val presentConfigMap: Map<String?, Map<Long, Any>> by lazy {
+        return@lazy if (Utils.IS_TIM) {
+            TIMConfigTable.configs
+        } else {
+            QQConfigTable.configs
         }
-        return QQConfigTable.getConfig(className) as T
+    }
+    private val presentRangeConfigMap: Map<String?, Map<Long, Any>> by lazy {
+        return@lazy if (Utils.IS_TIM) {
+            TIMConfigTable.rangingConfigs
+        } else {
+            QQConfigTable.rangingConfigs
+        }
+    }
+
+    private val cacheMap: Map<String?, Any?> by lazy {
+        val map: HashMap<String?, Any?> = HashMap()
+        val versionCode = Utils.getHostVersionCode()
+        for (pair in presentRangeConfigMap) {
+            for (i in versionCode downTo 1) {
+                if (pair.value.containsKey(i)) {
+                    map[pair.key] = pair.value[i]
+                    break
+                }
+            }
+        }
+        for (pair in presentConfigMap) {
+            if (pair.value.containsKey(versionCode)) {
+                map[pair.key] = pair.value[versionCode]
+            }
+        }
+        map
+    }
+
+    fun <T> getConfig(className: String?): T {
+        val config = cacheMap[className]
+        return config as T
+                ?: throw RuntimeException("$className :Unsupported Version: "+Utils.getHostVersionCode())
     }
 
 }
