+++ app/src/main/java/me/singleneuron/qn_kernel/tlb/TIMConfigTable.kt
@@ -4,15 +4,14 @@ import me.ketal.util.TIMVersion.*
 import nil.nadph.qnotified.bridge.QQMessageFacade
 import nil.nadph.qnotified.hook.MultiActionHook
 import nil.nadph.qnotified.hook.ReplyNoAtHook
-import nil.nadph.qnotified.util.Utils
 
 object TIMConfigTable {
 
-    private val configs: Map<String?, Map<Long, Any>> = mapOf(
+    val configs: Map<String?, Map<Long, Any>> = mapOf(
 
     )
 
-    private val rangingConfigs: Map<String?, Map<Long, Any>> = mapOf(
+    val rangingConfigs: Map<String?, Map<Long, Any>> = mapOf(
 
             MultiActionHook::class.java.simpleName to mapOf(
                     TIM_1_0_0 to "a",
@@ -35,30 +34,5 @@ object TIMConfigTable {
             ),
 
     )
-
-    private val cacheMap: Map<String?, Any?> by lazy {
-        val map: HashMap<String?, Any?> = HashMap()
-        val versionCode = Utils.getHostVersionCode()
-        for (pair in rangingConfigs) {
-            for (i in versionCode downTo TIM_1_0_0) {
-                if (pair.value.containsKey(i)) {
-                    map[pair.key] = pair.value[i]
-                    break
-                }
-            }
-        }
-        for (pair in configs) {
-            if (pair.value.containsKey(versionCode)) {
-                map[pair.key] = pair.value[versionCode]
-            }
-        }
-        map
-    }
-
-    fun <T> getConfig(className: String?): T {
-        val config = cacheMap[className]
-        return config as T
-                ?: throw RuntimeException("$className :Unsupported TIM Version")
-    }
 
 }
