+++ app/src/main/java/nil/nadph/qnotified/HookEntry.java
@@ -44,8 +44,9 @@ public class HookEntry implements IXposedHookLoadPackage {
             case PACKAGE_NAME_SELF:
                 XposedHelpers.findAndHookMethod("nil.nadph.qnotified.util.Utils", lpparam.classLoader, "getActiveModuleVersion", XC_MethodReplacement.returnConstant(Utils.QN_VERSION_NAME));
                 break;
-            case PACKAGE_NAME_QQ:
             case PACKAGE_NAME_TIM:
+                Utils.IS_TIM = true;
+            case PACKAGE_NAME_QQ:
                 StartupHook.getInstance().doInit(lpparam.classLoader);
                 break;
             case PACKAGE_NAME_QQ_INTERNATIONAL:
